Avoid warning.

+++ cmd-load-buffer.c
@@ -1,4 +1,4 @@
-/* $Id: cmd-load-buffer.c,v 1.1 2009-01-25 19:00:10 tcunha Exp $ */
+/* $Id: cmd-load-buffer.c,v 1.2 2009-01-27 23:26:15 nicm Exp $ */
 
 /*
  * Copyright (c) 2009 Tiago Cunha <me@tiagocunha.org>
@@ -77,7 +77,7 @@ cmd_load_buffer_exec(struct cmd *self, struct cmd_ctx *ctx)
 		return (-1);
 	}
 
-	if (fread(buf, 1, statbuf.st_size, f) != statbuf.st_size) {
+	if (fread(buf, 1, statbuf.st_size, f) != (size_t) statbuf.st_size) {
 		ctx->error(ctx, "%s: fread error", data->arg);
 		xfree(buf);
 		fclose(f);
